fix: harden watch decoding and rate limiting

+++ PaicordLib/Sources/DiscordModels/Types/Audit Log.swift
@@ -6,7 +6,7 @@ public struct AuditLog: Sendable, Codable {
 
     public enum Mixed: Sendable, Codable, CustomStringConvertible {
       case string(String)
-      case int(Int)
+      case int(Int64)
       case double(Double)
       case bool(Bool)
       case strings([String])
@@ -50,7 +50,7 @@ public struct AuditLog: Sendable, Codable {
         let container = try decoder.singleValueContainer()
         if let string = try? container.decode(String.self) {
           self = .string(string)
-        } else if let int = try? container.decode(Int.self) {
+        } else if let int = try? container.decode(Int64.self) {
           self = .int(int)
         } else if let bool = try? container.decode(Bool.self) {
           self = .bool(bool)
