+++ PaicordLib/Sources/DiscordModels/Types/Gateway+Payloads.swift
@@ -589,7 +589,7 @@ extension Gateway {
     public var nsfw: Bool
     public var application_command_counts: [String: Int]?
     public var embedded_activities: [Gateway.Activity]?
-    public var version: Int?
+    public var version: Int64?
     public var guild_id: GuildSnowflake?
     /// Extra fields:
     public var joined_at: DiscordTimestamp
@@ -807,7 +807,7 @@ extension Gateway {
   public struct GuildRoleDelete: Sendable, Codable {
     public var guild_id: GuildSnowflake
     public var role_id: RoleSnowflake
-    public var version: Int?
+    public var version: Int64?
   }
 
   /// Not the same as what Discord calls `Guild Scheduled Event User`.
@@ -2047,14 +2047,14 @@ extension Gateway {
     public var mention_count: Int?
     public var flags: IntBitField<ReadState.Flags>?
     public var last_viewed: Int?
-    public var version: Int
+    public var version: Int64
   }
 
   /// https://docs.discord.food/topics/gateway-events#channel-pins-ack
   public struct ChannelPinsAcknowledge: Sendable, Codable {
     public var channel_id: ChannelSnowflake
     public var timestamp: DiscordTimestamp
-    public var version: Int
+    public var version: Int64
   }
 
   /// https://docs.discord.food/topics/gateway-events#user-non-channel-ack-structure
@@ -2062,7 +2062,7 @@ extension Gateway {
     public var ack_type: ReadState.Kind
     public var resource_id: UserSnowflake
     public var entity_id: AnySnowflake
-    public var version: Int
+    public var version: Int64
   }
 
   /// https://docs.discord.food/resources/message#create-attachments
